change logo 	renamed:    redash/static/images/favicon-16x16.png -> client/app/assets/images/favicon-16x16.png 	renamed:    redash/static/images/favicon-32x32.png -> client/app/assets/images/favicon-32x32.png 	renamed:    redash/static/images/favicon-96x96.png -> client/app/assets/images/favicon-96x96.png 	renamed:    redash/static/images/logo_small.png -> client/app/assets/images/logo_small.png 	deleted:    client/app/assets/images/redash_icon_small.png 	modified:   client/app/components/app-header/index.js 	modified:   client/app/pages/dashboards/public-dashboard-page.js 	modified:   client/app/pages/queries/visualization-embed.js 	modified:   client/app/services/notifications.js 	deleted:    redash/static/images/logo.png
