+++ client/app/assets/images/logo_small.png
